Create separate applications table component
